+++ daily/2020/2020-05/2020-05-20/index2.ts
@@ -51,3 +51,5 @@ demo();
 // timeout(3000).then((value) => {
 //   console.log(value);
 // });
+
+export {};